refactor: zero clippy warnings at pedantic level; gate it in check.sh
- 41 format_push_string -> write!/writeln!; 4 format_collect -> fold+write!
- 13 assigning_clones -> clone_from; 12 manual_let_else -> let-else
- 8 match_same_arms merged; 9 items_after_statements moved to fn tops
- 6 needless_pass_by_value -> &str/&T (App::new key, copy_text,
  pick_model, save_clipboard_image, split_citation_span, office_text)
- 4 unnecessary_wraps dropped (picker_enter, confirm_files_add,
  new_session, confirm_space) + callers
- JSON tool-arg indices (openrouter x5, tools x7) now usize/u32/i32
  try_from with safe defaults; human_size takes u64 (sizes unsigned_abs)
- module-level cast allows with justification (terminal-bounded dims,
  bounded token/stat/list math); too_many_lines allows on long-by-design
  dispatch/orchestration fns; struct_field_names on App/Usage,
  struct_excessive_bools on App/Settings (postfixes and bools are meaning)
- unused_self (slug_from_ocr, descriptive_paste_name), unused_async
  (drain_steers), used_underscore_binding, unnecessary_debug_formatting,
  case_sensitive ZIP comparisons (allowed: spec-exact), default_trait_access
- similar_names renames (txt->body, argv->cmd); doc_markdown backticks
- check.sh now gates clippy -D warnings -W clippy::pedantic (was default
  lints only). Nursery's option_if_let_else deliberately ungated: the
  if-let-else form is more readable at the flagged sites

+++ src/ui/popups/images.rs
@@ -18,7 +18,7 @@ pub(crate) fn render(f: &mut Frame, app: &App) {
             let created = crate::ui::fmt_created(&img.modified);
             ListItem::new(Line::from(vec![
                 Span::styled(img.name.clone(), Style::default().fg(app.theme.fg)),
-                Span::styled(format!("  {}", crate::app::human_size(img.size as i64)), dim),
+                Span::styled(format!("  {}", crate::app::human_size(img.size)), dim),
                 Span::styled(format!("  {created}"), dim),
             ]))
         })
